Fix tests from previous commit's changes.

+++ examples/simple_input.rs
@@ -28,7 +28,7 @@ fn main() {
     {
         let mut fb = {
             // Create the frame buffer
-            let mut fb = FrameBufferMut::new(width as usize, height as usize);
+            let mut fb = FrameBufferMut::new(width, height);
             fb.insert_channels(&[("R", 0.0), ("G", 0.0), ("B", 0.0)], &mut pixel_data);
             fb
         };
